# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/projects/{id}/milestones": {
      "post": {
        "operationId": "ProjectsController_addMilestone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "el"
            },
            "description": "el or en. Greek when the header is absent."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titleEl": {
                    "type": "string"
                  },
                  "baselineDate": {
                    "type": "string"
                  },
                  "forecastDate": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "actualDate": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "isGate": {
                    "type": "boolean"
                  },
                  "sortOrder": {
                    "type": "integer",
                    "minimum": -9007199254740991,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "titleEl",
                  "baselineDate",
                  "isGate"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The milestone as stored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "projectId": {
                      "type": "string"
                    },
                    "titleEl": {
                      "type": "string"
                    },
                    "baselineDate": {
                      "type": "string"
                    },
                    "forecastDate": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "actualDate": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "isGate": {
                      "type": "boolean"
                    },
                    "sortOrder": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "projectId",
                    "titleEl",
                    "baselineDate",
                    "forecastDate",
                    "actualDate",
                    "isGate",
                    "sortOrder"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "The body is not a valid milestone",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "The i18n key of the message, for the client to re-translate."
                    },
                    "message": {
                      "type": "string",
                      "description": "The message in the caller's language; Greek unless asked."
                    },
                    "requestId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "key",
                    "message",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "A read-only account, or a role that does not run projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "The i18n key of the message, for the client to re-translate."
                    },
                    "message": {
                      "type": "string",
                      "description": "The message in the caller's language; Greek unless asked."
                    },
                    "requestId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "key",
                    "message",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "No such project, or none the caller may see",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "The i18n key of the message, for the client to re-translate."
                    },
                    "message": {
                      "type": "string",
                      "description": "The message in the caller's language; Greek unless asked."
                    },
                    "requestId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "key",
                    "message",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a milestone to the project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{id}/milestones/{mid}": {
      "patch": {
        "operationId": "ProjectsController_updateMilestone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "mid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "el"
            },
            "description": "el or en. Greek when the header is absent."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titleEl": {
                    "type": "string"
                  },
                  "baselineDate": {
                    "type": "string"
                  },
                  "forecastDate": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "actualDate": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "isGate": {
                    "type": "boolean"
                  },
                  "sortOrder": {
                    "type": "integer",
                    "minimum": -9007199254740991,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "titleEl",
                  "baselineDate",
                  "isGate"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The milestone after the change",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "projectId": {
                      "type": "string"
                    },
                    "titleEl": {
                      "type": "string"
                    },
                    "baselineDate": {
                      "type": "string"
                    },
                    "forecastDate": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "actualDate": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "isGate": {
                      "type": "boolean"
                    },
                    "sortOrder": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "projectId",
                    "titleEl",
                    "baselineDate",
                    "forecastDate",
                    "actualDate",
                    "isGate",
                    "sortOrder"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "The body is not a valid milestone",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "The i18n key of the message, for the client to re-translate."
                    },
                    "message": {
                      "type": "string",
                      "description": "The message in the caller's language; Greek unless asked."
                    },
                    "requestId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "key",
                    "message",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "A read-only account, or a role that does not run projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "The i18n key of the message, for the client to re-translate."
                    },
                    "message": {
                      "type": "string",
                      "description": "The message in the caller's language; Greek unless asked."
                    },
                    "requestId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "key",
                    "message",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "No such project or milestone",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "The i18n key of the message, for the client to re-translate."
                    },
                    "message": {
                      "type": "string",
                      "description": "The message in the caller's language; Greek unless asked."
                    },
                    "requestId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "key",
                    "message",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "The body asks to move the baseline date",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "The i18n key of the message, for the client to re-translate."
                    },
                    "message": {
                      "type": "string",
                      "description": "The message in the caller's language; Greek unless asked."
                    },
                    "requestId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "key",
                    "message",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a milestone; the baseline date cannot move",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "eCapital API",
    "description": "Capital projects and maintenance for ΟΚΥπΥ. M0: sign in, see your own org units and their area tree, read the audit log. Error messages come back in Greek unless Accept-Language asks for English.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Entra ID access token, or a development token while DEV_AUTH is on."
      }
    },
...
  }
}